Refactor model helper function to its own file
