+++ codegen/smithy-aws-kotlin-codegen/src/test/kotlin/aws/sdk/kotlin/codegen/testutil/ModelUtil.kt
@@ -1,0 +1,16 @@
+package aws.sdk.kotlin.codegen.testutil
+
+import software.amazon.smithy.kotlin.codegen.test.prependNamespaceAndService
+import software.amazon.smithy.kotlin.codegen.test.toSmithyModel
+import software.amazon.smithy.model.Model
+
+internal fun model(serviceName: String): Model =
+    """
+        @http(method: "PUT", uri: "/foo")
+        operation Foo { }
+        
+        @http(method: "POST", uri: "/bar")
+        operation Bar { }
+    """
+        .prependNamespaceAndService(operations = listOf("Foo", "Bar"), serviceName = serviceName)
+        .toSmithyModel()
